feat(workers): makes workers more modular by abstracting the base genmain functions
